Hackverse/AYUR-PROJECT/backend/requirements.txt
flask
flask-cors
opencv-python
numpy
torch
torchvision
pillow
pandas
scikit-learn
matplotlib
joblib
gunicorn
